リファクタリング: ショップ/宿屋ロジックを shop.py に集約（party_manager 分割の仕上げ）
「ショップ処理がどこにあるか」の答えが party_manager / shop.py / shop_service の
3箇所に割れていた問題を解消。購入・宿泊の本体を、未使用化していた shop.py に集約し
正規のショップ/宿屋ドメインとして再生した。

- buy_item / buy_monster / rest_at_inn を party_manager から shop.py へ移動。
- shop.py: player を import すると循環するため、Player/Location 参照を
  TYPE_CHECKING 化（runtime 依存ゼロ）。buy_monster は player ファサード経由で
  add_monster_to_party を呼ぶよう変更。
- Player ファサードの shop 系委譲先を shop へ（公開メソッド名は不変）。
- party_manager は隊列・アイテムだけの責務に縮小し、未使用 import も整理。

これで party_manager（隊列）/ shop（店・宿）/ equipment_manager（装備）/
synthesis_manager（合成）と、ドメインが概念ごとに分かれた。テスト85件グリーン。
