+++ backend/src/monster_rpg/party_manager.py
@@ -6,7 +6,6 @@ from typing import Optional
 
 from .monsters.monster_class import Monster
 from .monsters.monster_data import ALL_MONSTERS
-from .items.item_data import ALL_ITEMS
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -122,44 +121,3 @@ def use_item(player: "Player", item_idx: int, target_monster: Monster) -> bool:
     if success:
         player.items.pop(item_idx)
     return success
-
-
-def rest_at_inn(player: "Player", cost: int) -> bool:
-    if player.gold >= cost:
-        player.gold -= cost
-        print(f"{cost}G を支払い、宿屋に泊まった。")
-        for monster in player.party_monsters:
-            monster.hp = monster.max_hp
-            monster.mp = monster.max_mp
-            monster.is_alive = True
-        print("パーティは完全に回復した！")
-        return True
-    else:
-        print("お金が足りない！宿屋に泊まれない...")
-        return False
-
-
-def buy_item(player: "Player", item_id: str, price: int) -> bool:
-    if player.gold < price:
-        print("お金が足りない！")
-        return False
-    if item_id not in ALL_ITEMS:
-        print("そのアイテムは存在しない。")
-        return False
-    player.gold -= price
-    player.items.append(ALL_ITEMS[item_id])
-    print(f"{ALL_ITEMS[item_id].name} を {price}G で購入した。")
-    return True
-
-
-def buy_monster(player: "Player", monster_id: str, price: int) -> bool:
-    if player.gold < price:
-        print("お金が足りない！")
-        return False
-    if monster_id not in ALL_MONSTERS:
-        print("そのモンスターは存在しない。")
-        return False
-    player.gold -= price
-    add_monster_to_party(player, monster_id)
-    print(f"{ALL_MONSTERS[monster_id].name} を {price}G で仲間にした。")
-    return True
